+++ bindings/nostr-sdk-js/src/lib.rs
@@ -20,7 +20,6 @@ pub mod duration;
 pub mod error;
 pub mod logger;
 pub mod nwc;
-pub mod pool;
 pub mod profile;
 pub mod protocol;
 pub mod relay;